CFIN-252 Refactor so an error log reports the same initial information as an audit log, so all requests captured. * auditLogEntry renamed to be logEntry * added LOG_TYPES constant
