+++ src/constants/constants.js
@@ -1,0 +1,5 @@
+const LOG_TYPES = {
+    ERROR: "error",
+    AUDIT: "audit",
+};
+module.exports = { LOG_TYPES };
